backend/requirements.txt
fastapi
uvicorn
pydantic
transformers
psycopg2-binary
python-dotenv